Add ability to specify path parameters via flags in the client CLI.
Both these can be used now:

```
cellar-cli list bottles /cellar/accounts/1/bottles --years=2012
cellar-cli list bottles --accountsID=1 --years=2012
```

The path takes precedence when specified.
