+++ goagen/gen_client/generator_test.go
@@ -4,14 +4,12 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/goadesign/goa/design"
 	"github.com/goadesign/goa/goagen/codegen"
 	"github.com/goadesign/goa/goagen/gen_client"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/onsi/gomega/gexec"
 )
 
 var _ = Describe("Generate", func() {
@@ -40,110 +38,6 @@ var _ = Describe("Generate", func() {
 	AfterEach(func() {
 		codegen.CommandName = oldCommand
 		os.RemoveAll(outDir)
-	})
-
-	Context("with a dummy API", func() {
-		BeforeEach(func() {
-			design.Design = &design.APIDefinition{
-				Name:        "testapi",
-				Title:       "dummy API with no resource",
-				Description: "I told you it's dummy",
-			}
-		})
-
-		It("generates a dummy app", func() {
-			Ω(genErr).Should(BeNil())
-			Ω(files).Should(HaveLen(5))
-			content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "testapi-cli", "main.go"))
-			Ω(err).ShouldNot(HaveOccurred())
-			Ω(len(strings.Split(string(content), "\n"))).Should(BeNumerically(">=", 16))
-			_, err = gexec.Build(filepath.Join(testgenPackagePath, "client", "testapi-cli"))
-			Ω(err).ShouldNot(HaveOccurred())
-		})
-	})
-
-	Context("with an action with an integer parameter with no default value", func() {
-		BeforeEach(func() {
-			codegen.TempCount = 0
-			design.Design = &design.APIDefinition{
-				Name:        "testapi",
-				Title:       "dummy API with no resource",
-				Description: "I told you it's dummy",
-				Resources: map[string]*design.ResourceDefinition{
-					"foo": {
-						Name: "foo",
-						Actions: map[string]*design.ActionDefinition{
-							"show": {
-								Name: "show",
-								QueryParams: &design.AttributeDefinition{
-									Type: design.Object{
-										"param": &design.AttributeDefinition{Type: design.Integer},
-										"time":  &design.AttributeDefinition{Type: design.DateTime},
-									},
-								},
-								Routes: []*design.RouteDefinition{
-									{
-										Verb: "GET",
-										Path: "",
-									},
-								},
-							},
-						},
-					},
-				},
-			}
-			fooRes := design.Design.Resources["foo"]
-			showAct := fooRes.Actions["show"]
-			showAct.Parent = fooRes
-			showAct.Routes[0].Parent = showAct
-		})
-
-		It("generates the correct command flag initialization code", func() {
-			Ω(genErr).Should(BeNil())
-			Ω(files).Should(HaveLen(6))
-			content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "testapi-cli", "commands.go"))
-			Ω(err).ShouldNot(HaveOccurred())
-			Ω(len(strings.Split(string(content), "\n"))).Should(BeNumerically(">=", 3))
-			Ω(content).Should(ContainSubstring("var tmp2 int"))
-			Ω(content).Should(ContainSubstring("var tmp3 string"))
-			Ω(content).Should(ContainSubstring(".Flags()"))
-			_, err = gexec.Build(filepath.Join(testgenPackagePath, "client", "testapi-cli"))
-			Ω(err).ShouldNot(HaveOccurred())
-
-		})
-
-		Context("with an action with a multiline description", func() {
-			const multiline = "multi\nline"
-
-			BeforeEach(func() {
-				design.Design.Resources["foo"].Actions["show"].Description = multiline
-			})
-
-			It("properly escapes the multi-line string used in the short description", func() {
-				Ω(genErr).Should(BeNil())
-				Ω(files).Should(HaveLen(6))
-				content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "testapi-cli", "main.go"))
-				Ω(err).ShouldNot(HaveOccurred())
-				Ω(string(content)).Should(ContainSubstring(multiline))
-			})
-		})
-
-		Context("with an action with a description containing backticks", func() {
-			const pre = "pre"
-			const post = "post"
-
-			BeforeEach(func() {
-				design.Design.Resources["foo"].Actions["show"].Description = pre + "`" + post
-			})
-
-			It("properly escapes the multi-line string used in the short description", func() {
-				Ω(genErr).Should(BeNil())
-				Ω(files).Should(HaveLen(6))
-				content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "testapi-cli", "main.go"))
-				Ω(err).ShouldNot(HaveOccurred())
-				Ω(string(content)).Should(ContainSubstring(pre + "` + \"`\" + `" + post))
-			})
-		})
 	})
 
 	Context("with an action with security configured", func() {
@@ -207,14 +101,6 @@ var _ = Describe("Generate", func() {
 			content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "foo.go"))
 			Ω(err).ShouldNot(HaveOccurred())
 			Ω(content).Should(ContainSubstring("c.SignerJWT1.Sign(ctx, req)"))
-		})
-
-		It("generates the Signer.RegisterFlags call from Command", func() {
-			Ω(genErr).Should(BeNil())
-			Ω(files).Should(HaveLen(6))
-			content, err := ioutil.ReadFile(filepath.Join(outDir, "client", "testapi-cli", "commands.go"))
-			Ω(err).ShouldNot(HaveOccurred())
-			Ω(content).Should(ContainSubstring("c.SignerJWT1.RegisterFlags(cc)"))
 		})
 	})
 })
